requirements.txt
selenium ~= 4.1.0
opencv-python ~= 4.5.5.62
pillow ~= 9.0.0
scikit-image